chore: move utils into own folder
